refactor: refactored previous tasks
